mb/lenovo/s230u: Fix compiling for 64bit mode
This fixes the warning when an integer is cast to a pointer of a
different size.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/81381
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/mainboard/lenovo/s230u/ec.h
@@ -5,7 +5,7 @@
 
 void lenovo_s230u_ec_init(void);
 
-#define ECMM(x) (*((volatile u8 *)(CONFIG_EC_BASE_ADDRESS + x)))
+#define ECMM(x) (*((volatile u8 *)(uintptr_t)(CONFIG_EC_BASE_ADDRESS + x)))
 #define ec_mm_read(addr) (ECMM(0x100 + addr))
 #define ec_mm_write(addr, data) (ECMM(0x100 + addr) = data)
 #define ec_mm_set_bit(addr, bit) (ECMM(0x100 + addr) |= 1 << bit)
